[WNMGDS-760] Convert DateField and ChoiceList components to typescript (#1041)
