+++ packages/design-system-docs/src/pages/components/ChoiceList/_ChoiceList.docs.scss
@@ -17,7 +17,7 @@ A `ChoiceList` component can be used to render a radio button group, or checkbox
 
 @react-example ChoiceList.example.jsx
 
-@react-props ChoiceList.jsx
+@react-props ChoiceList.tsx
 
 Style guide: components.choice.choicelist
 */
@@ -31,7 +31,7 @@ Any _undocumented_ props that you pass to this component will be passed to the `
 
 @react-example Choice.example.jsx
 
-@react-props Choice.jsx
+@react-props Choice.tsx
 
 Style guide: components.choice.react
 */
