+++ packages/design-system-docs/src/pages/components/DateField/_DateField.docs.scss
@@ -15,7 +15,7 @@ Style guide: components.date-field
 
 @react-example DateField.example.jsx
 
-@react-props DateField.jsx
+@react-props DateField.tsx
 
 Style guide: components.date-field.react
 */
