+++ packages/design-system/src/components/DateField/DateField.test.tsx
@@ -1,7 +1,8 @@
 jest.mock('lodash.uniqueid', () => (str) => `${str}snapshot`);
 /* eslint-disable import/first */
-import { DateField, defaultDateFormatter } from './DateField';
+import { DateField } from './DateField';
 import React from 'react';
+import defaultDateFormatter from './defaultDateFormatter';
 import renderer from 'react-test-renderer';
 
 describe('DateField', () => {
