+++ packages/design-system/src/types/index.d.ts
@@ -6,10 +6,6 @@ export { default as Alert } from './Alert/Alert';
 export { default as Autocomplete } from './Autocomplete/Autocomplete';
 export { default as Badge } from './Badge/Badge';
 export { default as Button } from './Button/Button';
-export { default as Choice } from './ChoiceList/Choice';
-export { default as ChoiceList } from './ChoiceList/ChoiceList';
-export { default as DateField } from './DateField/DateField';
-export { default as DateInput } from './DateField/DateInput';
 export { default as Dialog } from './Dialog/Dialog';
 export { default as Dropdown } from './Dropdown/Dropdown';
 export { default as Select } from './Dropdown/Select';
@@ -38,6 +34,10 @@ export { default as UsaBanner } from './UsaBanner/UsaBanner';
 export { default as VerticalNav } from './VerticalNav/VerticalNav';
 
 // Type definitions generated by cmsds scripts
+export * from './ChoiceList/Choice';
+export * from './ChoiceList/ChoiceList';
+export * from './DateField/DateField';
+export * from './DateField/DateInput';
 export * from './InlineError/InlineError';
 export * from './FilterChip/FilterChip';
 export * from './FormControl/FormControl';
